DEV: move api-runner to lib
