+++ libs/shared/data-access/api-runner/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/api-runner';
 export * from './lib/shared-data-access-api-runner.module';
